requirements.txt
pandas==1.3.5
Pillow==9.1.1
